TIP-303: move StepFactory from BatchBundle to component
