+++ spec/Akeneo/Bundle/BatchBundle/Connector/ConnectorRegistrySpec.php
@@ -3,7 +3,7 @@
 namespace spec\Akeneo\Bundle\BatchBundle\Connector;
 
 use Akeneo\Component\Batch\Job\JobFactory;
-use Akeneo\Bundle\BatchBundle\Step\StepFactory;
+use Akeneo\Component\Batch\Step\StepFactory;
 use Akeneo\Component\Batch\Job\Job;
 use Akeneo\Component\Batch\Model\JobInstance;
 use Akeneo\Component\Batch\Step\ItemStep;
